requirements.txt
# Milvus 多路检索验证项目依赖

# 核心依赖
pymilvus>=2.4.0
elasticsearch>=8.0.0
zhipuai>=2.0.0

# 文本处理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2
jieba>=0.42.1

# 向量和稀疏处理
scipy>=1.11.0

# 数据处理
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# HTTP 请求
requests>=2.31.0
httpx>=0.24.0

# 日志和进度
loguru>=0.7.0
tqdm>=4.65.0

# 可视化（可选）
matplotlib>=3.7.0
seaborn>=0.12.0

# 开发工具
jupyter>=1.0.0
pytest>=7.4.0
black>=23.0.0
